fix: PR Comment 반영
Related: #39
