+++ src/main/java/com/example/dgbackend/domain/combination/controller/CombinationController.java
@@ -18,7 +18,7 @@ import java.util.List;
 
 @Tag(name = "오늘의 조합 API")
 @RestController
-@RequestMapping("/api/combinations")
+@RequestMapping("/combinations")
 @RequiredArgsConstructor
 public class CombinationController {
 
@@ -70,17 +70,6 @@ public class CombinationController {
         return ApiResponse.onSuccess(combinationQueryService.getCombinationEditDTO(combinationId));
     }
 
-//    @Operation(summary = "오늘의 조합 수정", description = "특정 오늘의 조합을 수정합니다.")
-//    @ApiResponses(value = {
-//            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "오늘의 조합 수정 성공")
-//    })
-//    @Parameter(name = "combinationId", description = "오늘의 조합 Id, Path Variable 입니다.")
-//    @PatchMapping("/{combinationId}")
-//    public ApiResponse<> editCombination(@PathVariable(name = "combinationId") Long combinationId) {
-//
-//        return null;
-//    }
-//
     @Operation(summary = "오늘의 조합 삭제", description = "특정 오늘의 조합을 삭제합니다.")
     @ApiResponses(value = {
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "오늘의 조합 삭제 성공")
