+++ src/main/java/com/example/dgbackend/domain/combination/domain/Combination.java
@@ -2,8 +2,8 @@ package com.example.dgbackend.domain.combination.domain;
 
 
 import com.example.dgbackend.domain.combinationcomment.domain.CombinationComment;
-import com.example.dgbackend.domain.combinationimage.CombinationImage;
-import com.example.dgbackend.domain.member.Member;
+import com.example.dgbackend.domain.combinationimage.domain.CombinationImage;
+import com.example.dgbackend.domain.member.domain.Member;
 import com.example.dgbackend.global.common.BaseTimeEntity;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
