+++ src/main/java/com/example/dgbackend/domain/combination/dto/CombinationResponse.java
@@ -2,9 +2,9 @@ package com.example.dgbackend.domain.combination.dto;
 
 import com.example.dgbackend.domain.combination.domain.Combination;
 import com.example.dgbackend.domain.combinationcomment.domain.CombinationComment;
-import com.example.dgbackend.domain.combinationimage.CombinationImage;
-import com.example.dgbackend.domain.hashtagoption.HashTagOption;
-import com.example.dgbackend.domain.member.Member;
+import com.example.dgbackend.domain.combinationimage.domain.CombinationImage;
+import com.example.dgbackend.domain.hashtagoption.domain.HashTagOption;
+import com.example.dgbackend.domain.member.domain.Member;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
