+++ src/main/java/com/example/dgbackend/domain/combinationimage/domain/CombinationImage.java
@@ -1,6 +1,6 @@
-package com.example.dgbackend.domain.recommend;
+package com.example.dgbackend.domain.combinationimage.domain;
 
-import com.example.dgbackend.domain.member.Member;
+import com.example.dgbackend.domain.combination.domain.Combination;
 import com.example.dgbackend.global.common.BaseTimeEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -10,42 +10,24 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
+@Builder
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class Recommend extends BaseTimeEntity {
+public class CombinationImage extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @NotNull
-    private Integer desireLevel;
-
-    @NotNull
-    private String foodName;
-
-    private String feeling;
-
-    private String weather;
-
-    @NotNull
-    private String drinkName;
-
-    @NotNull
-    private String drinkInfo;
-
-    @NotNull
     private String imageUrl;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
+    @JoinColumn(name = "combination_id")
+    private Combination combination;
 
 }
