+++ src/main/java/com/example/dgbackend/domain/hashtagoption/repository/HashTagOptionRepository.java
@@ -1,7 +1,7 @@
 package com.example.dgbackend.domain.hashtagoption.repository;
 
 import com.example.dgbackend.domain.combination.domain.Combination;
-import com.example.dgbackend.domain.hashtagoption.HashTagOption;
+import com.example.dgbackend.domain.hashtagoption.domain.HashTagOption;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
