+++ src/main/java/com/example/dgbackend/domain/recommend/domain/Recommend.java
@@ -1,6 +1,6 @@
-package com.example.dgbackend.domain.combinationimage;
+package com.example.dgbackend.domain.recommend.domain;
 
-import com.example.dgbackend.domain.combination.Combination;
+import com.example.dgbackend.domain.member.domain.Member;
 import com.example.dgbackend.global.common.BaseTimeEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -19,17 +19,33 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class CombinationImage extends BaseTimeEntity {
+public class Recommend extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @NotNull
+    private Integer desireLevel;
+
+    @NotNull
+    private String foodName;
+
+    private String feeling;
+
+    private String weather;
+
+    @NotNull
+    private String drinkName;
+
+    @NotNull
+    private String drinkInfo;
+
+    @NotNull
     private String imageUrl;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "combination_id")
-    private Combination combination;
+    @JoinColumn(name = "member_id")
+    private Member member;
 
 }
